Change if statement to try/except block
This helps to avoid errors because of race conditions of jobs in the
cluster. Another job may create the directory after
the current job has has checked that the directory doesn't exist, but
before it has created the directory, causing a FileExistsError.

See "look before you leap" vs "easier to ask for forgiveness than
permission".